+++ packages/flow-sdk/src/core/index.ts
@@ -1,4 +1,4 @@
 export * from './client';
 export * from './chains';
 export * from './wallet';
-export * from './wallet-utilities';
+export * from './utils';
