+++ packages/flow-sdk/src/core/utils/types.test.ts
@@ -6,7 +6,7 @@ import {
   type SocialProvider
 } from './types';
 
-describe('Wallet Utilities Types', () => {
+describe('Utils Types', () => {
   describe('AccountBalanceParams', () => {
     it('should accept valid parameters', () => {
       const mockClient = {} as FlowClient;
