--- packages/flow-sdk/src/core/wallet-utilities/index.ts
@@ -1,2 +1,0 @@
-export * from './wallet-utilities';
-export * from './types';
